requirements.txt
beautifulsoup4==4.10.0
black==21.12b0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
idna==3.3
lxml==4.7.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
spotipy==2.19.0
tomli==1.2.3
typing-extensions==4.0.1
urllib3==1.26.8
